Move all deprecated functions to deprecated.py
