+++ pysr/__init__.py
@@ -1,9 +1,10 @@
 from . import sklearn_monkeypatch
+from .deprecated import best, best_callable, best_row, best_tex, pysr
 from .export_jax import sympy2jax
 from .export_torch import sympy2torch
 from .feynman_problems import FeynmanProblem, Problem
 from .julia_helpers import install
-from .sr import PySRRegressor, best, best_callable, best_row, best_tex, pysr
+from .sr import PySRRegressor
 from .version import __version__
 
 __all__ = [
